services/ai_core/requirements.txt
fastapi
uvicorn[standard]
langchain-google-genai
langchain
langgraph
asyncpg
redis
celery
httpx
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
cryptography
python-multipart
tenacity
pandas
PyJWT
zeep
redis[hiredis]
